fix old and new animations
