+++ source/js/modules/menu.js
@@ -1,12 +1,9 @@
+import updateImg from './updateImg';
+
 export default () => {
   let header = document.querySelector(`.js-header`);
   let menuToggler = document.querySelector(`.js-menu-toggler`);
   let menuLinks = document.querySelectorAll(`.js-menu-link`);
-
-  const updateImg = (selector, newSrc) => {
-    const imgWrap = document.querySelector(selector);
-    imgWrap.src = newSrc;
-  };
 
   if (menuToggler) {
     menuToggler.addEventListener(`click`, function () {
@@ -27,9 +24,7 @@ export default () => {
         document.body.classList.remove(`menu-opened`);
       }
       if (menuLinks[i].classList.contains(`prizes-link`)) {
-        updateImg(`.primary-award`, `img/module-3/primary-award.svg?` + new Date().getTime());
-        updateImg(`.secondary-award`, `img/module-3/secondary-award.svg?` + new Date().getTime());
-        updateImg(`.additional-award`, `img/module-3/additional-award.svg?` + new Date().getTime());
+        updateImg();
       }
     });
   }
